+++ pkg/codegen/go/gen.go
@@ -707,6 +707,51 @@ func (pkg *pkgContext) genEnumInputFuncs(w io.Writer, typeName string, enum *sch
 	fmt.Fprintf(w, "return %[1]s(e).To%[2]sOutputWithContext(ctx).To%[2]sPtrOutputWithContext(ctx)\n", elementType, asFuncName)
 	fmt.Fprintln(w, "}")
 	fmt.Fprintln(w)
+
+	details := pkg.detailsForEnum(enum)
+	// Generate the array input.
+	if details.arrayElement {
+		genInputInterface(w, typeName+"Array")
+
+		fmt.Fprintf(w, "type %[1]sArray []%[1]s\n\n", typeName)
+
+		genInputMethods(w, typeName+"Array", typeName+"Array", "[]"+typeName, false, false)
+	}
+
+	// Generate the map input.
+	if details.mapElement {
+		genInputInterface(w, typeName+"Map")
+
+		fmt.Fprintf(w, "type %[1]sMap map[string]%[1]s\n\n", typeName)
+
+		genInputMethods(w, typeName+"Map", typeName+"Map", "map[string]"+typeName, false, false)
+	}
+
+	// Generate the array output
+	if details.arrayElement {
+		fmt.Fprintf(w, "type %sArrayOutput struct { *pulumi.OutputState }\n\n", typeName)
+
+		genOutputMethods(w, typeName+"Array", "[]"+typeName, false)
+
+		fmt.Fprintf(w, "func (o %[1]sArrayOutput) Index(i pulumi.IntInput) %[2]sOutput {\n", typeName, elementType)
+		fmt.Fprintf(w, "\treturn pulumi.All(o, i).ApplyT(func (vs []interface{}) %sOutput {\n", elementType)
+		fmt.Fprintf(w, "\t\treturn vs[0].([]%s)[vs[1].(int)].To%sOutput()\n", typeName, asFuncName)
+		fmt.Fprintf(w, "\t}).(%sOutput)\n", elementType)
+		fmt.Fprintf(w, "}\n\n")
+	}
+
+	// Generate the map output.
+	if details.mapElement {
+		fmt.Fprintf(w, "type %sMapOutput struct { *pulumi.OutputState }\n\n", typeName)
+
+		genOutputMethods(w, typeName+"Map", "map[string]"+typeName, false)
+
+		fmt.Fprintf(w, "func (o %[1]sMapOutput) MapIndex(k pulumi.StringInput) %[2]sOutput {\n", typeName, elementType)
+		fmt.Fprintf(w, "\treturn pulumi.All(o, k).ApplyT(func (vs []interface{}) %sOutput {\n", elementType)
+		fmt.Fprintf(w, "\t\treturn vs[0].(map[string]%s)[vs[1].(string)].To%sOutput()\n", typeName, asFuncName)
+		fmt.Fprintf(w, "\t}).(%sOutput)\n", elementType)
+		fmt.Fprintf(w, "}\n\n")
+	}
 }
 
 func (pkg *pkgContext) genPlainType(w io.Writer, name, comment, deprecationMessage string,
@@ -731,7 +776,11 @@ func (pkg *pkgContext) genInputTypes(w io.Writer, t *schema.ObjectType, details 
 	fmt.Fprintf(w, "type %sArgs struct {\n", name)
 	for _, p := range t.Properties {
 		printCommentWithDeprecationMessage(w, p.Comment, p.DeprecationMessage, true)
-		fmt.Fprintf(w, "\t%s %s `pulumi:\"%s\"`\n", Title(p.Name), pkg.inputType(p.Type, !p.IsRequired), p.Name)
+		typ := pkg.inputType(p.Type, !p.IsRequired)
+		if p.IsPlain {
+			typ = pkg.plainType(p.Type, !p.IsRequired)
+		}
+		fmt.Fprintf(w, "\t%s %s `pulumi:\"%s\"`\n", Title(p.Name), typ, p.Name)
 	}
 	fmt.Fprintf(w, "}\n\n")
 
@@ -1011,7 +1060,7 @@ func (pkg *pkgContext) genResource(w io.Writer, r *schema.Resource, generateReso
 		case *schema.EnumType:
 			// not a pointer type and already handled above
 		default:
-			if p.IsRequired {
+			if p.IsRequired && !p.IsPlain {
 				fmt.Fprintf(w, "\tif args.%s == nil {\n", Title(p.Name))
 				fmt.Fprintf(w, "\t\treturn nil, errors.New(\"invalid value for required argument '%s'\")\n", Title(p.Name))
 				fmt.Fprintf(w, "\t}\n")
@@ -1167,7 +1216,11 @@ func (pkg *pkgContext) genResource(w io.Writer, r *schema.Resource, generateReso
 	fmt.Fprintf(w, "type %sArgs struct {\n", name)
 	for _, p := range r.InputProperties {
 		printCommentWithDeprecationMessage(w, p.Comment, p.DeprecationMessage, true)
-		fmt.Fprintf(w, "\t%s %s\n", Title(p.Name), pkg.inputType(p.Type, !p.IsRequired))
+		typ := pkg.inputType(p.Type, !p.IsRequired)
+		if p.IsPlain {
+			typ = pkg.plainType(p.Type, !p.IsRequired)
+		}
+		fmt.Fprintf(w, "\t%s %s\n", Title(p.Name), typ)
 	}
 	fmt.Fprintf(w, "}\n\n")
 
@@ -1400,6 +1453,12 @@ func (pkg *pkgContext) getTypeImports(t schema.Type, recurse bool, importsAndAli
 	}
 	seen[t] = struct{}{}
 	switch t := t.(type) {
+	case *schema.EnumType:
+		mod := pkg.tokenToPackage(t.Token)
+		if mod != pkg.mod {
+			p := path.Join(pkg.importBasePath, mod)
+			importsAndAliases[path.Join(pkg.importBasePath, mod)] = pkg.pkgImportAliases[p]
+		}
 	case *schema.ArrayType:
 		pkg.getTypeImports(t.ElementType, recurse, importsAndAliases, seen)
 	case *schema.MapType:
@@ -1757,24 +1816,44 @@ func generatePackageContextMap(tool string, pkg *schema.Package, goInfo GoPackag
 	// In addition, if the optional property's type is itself an object type, we also need to generate pointer
 	// types corresponding to all of it's nested properties, as our accessor methods will lift `nil` into
 	// those nested types.
-	var markOptionalPropertyTypesAsRequiringPtr func(seen codegen.StringSet, props []*schema.Property, parentOptional bool)
-	markOptionalPropertyTypesAsRequiringPtr = func(seen codegen.StringSet, props []*schema.Property, parentOptional bool) {
+	var populateDetailsForPropertyTypes func(seen codegen.StringSet, props []*schema.Property, parentOptional bool)
+	populateDetailsForPropertyTypes = func(seen codegen.StringSet, props []*schema.Property, parentOptional bool) {
 		for _, p := range props {
-			if obj, ok := p.Type.(*schema.ObjectType); ok && (!p.IsRequired || parentOptional) {
-				if seen.Has(obj.Token) {
-					continue
-				}
+			switch typ := p.Type.(type) {
+			case *schema.ObjectType:
+				if !p.IsRequired || parentOptional {
+					if seen.Has(typ.Token) {
+						continue
+					}
 
-				seen.Add(obj.Token)
-				getPkgFromToken(obj.Token).detailsForType(obj).ptrElement = true
-				markOptionalPropertyTypesAsRequiringPtr(seen, obj.Properties, true)
-			}
-			if enum, ok := p.Type.(*schema.EnumType); ok && (!p.IsRequired || parentOptional) {
-				if seen.Has(enum.Token) {
+					seen.Add(typ.Token)
+					getPkgFromToken(typ.Token).detailsForType(typ).ptrElement = true
+					populateDetailsForPropertyTypes(seen, typ.Properties, true)
+				}
+			case *schema.EnumType:
+				if seen.Has(typ.Token) {
 					continue
 				}
-				seen.Add(enum.Token)
-				getPkgFromToken(enum.Token).detailsForEnum(enum).ptrElement = true
+				if !p.IsRequired || parentOptional {
+					seen.Add(typ.Token)
+					getPkgFromToken(typ.Token).detailsForEnum(typ).ptrElement = true
+				}
+			case *schema.ArrayType:
+				if seen.Has(typ.String()) {
+					continue
+				}
+				seen.Add(typ.String())
+				if enum, ok := typ.ElementType.(*schema.EnumType); ok {
+					getPkgFromToken(enum.Token).detailsForEnum(enum).arrayElement = true
+				}
+			case *schema.MapType:
+				if seen.Has(typ.String()) {
+					continue
+				}
+				seen.Add(typ.String())
+				if enum, ok := typ.ElementType.(*schema.EnumType); ok {
+					getPkgFromToken(enum.Token).detailsForEnum(enum).mapElement = true
+				}
 			}
 		}
 	}
@@ -1789,14 +1868,20 @@ func generatePackageContextMap(tool string, pkg *schema.Package, goInfo GoPackag
 			if obj, ok := typ.ElementType.(*schema.ObjectType); ok {
 				getPkgFromToken(obj.Token).detailsForType(obj).arrayElement = true
 			}
+			if enum, ok := typ.ElementType.(*schema.EnumType); ok {
+				getPkgFromToken(enum.Token).detailsForEnum(enum).arrayElement = true
+			}
 		case *schema.MapType:
 			if obj, ok := typ.ElementType.(*schema.ObjectType); ok {
 				getPkgFromToken(obj.Token).detailsForType(obj).mapElement = true
 			}
+			if enum, ok := typ.ElementType.(*schema.EnumType); ok {
+				getPkgFromToken(enum.Token).detailsForEnum(enum).mapElement = true
+			}
 		case *schema.ObjectType:
 			pkg := getPkgFromToken(typ.Token)
 			pkg.types = append(pkg.types, typ)
-			markOptionalPropertyTypesAsRequiringPtr(seenMap, typ.Properties, false)
+			populateDetailsForPropertyTypes(seenMap, typ.Properties, false)
 		case *schema.EnumType:
 			pkg := getPkgFromToken(typ.Token)
 			pkg.enums = append(pkg.enums, typ)
@@ -1819,8 +1904,8 @@ func generatePackageContextMap(tool string, pkg *schema.Package, goInfo GoPackag
 			pkg.names.Add("Get" + resourceName(r))
 		}
 
-		markOptionalPropertyTypesAsRequiringPtr(seenMap, r.InputProperties, !r.IsProvider)
-		markOptionalPropertyTypesAsRequiringPtr(seenMap, r.Properties, !r.IsProvider)
+		populateDetailsForPropertyTypes(seenMap, r.InputProperties, !r.IsProvider)
+		populateDetailsForPropertyTypes(seenMap, r.Properties, !r.IsProvider)
 	}
 
 	scanResource(pkg.Provider)