+++ pkg/codegen/internal/test/testdata/simple-enum-schema/nodejs/tree/v1/index.ts
@@ -5,18 +5,22 @@ import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../../utilities";
 
 // Export members:
+export * from "./nursery";
 export * from "./rubberTree";
 
 // Export enums:
 export * from "../../types/enums/tree/v1";
 
 // Import resources to register:
+import { Nursery } from "./nursery";
 import { RubberTree } from "./rubberTree";
 
 const _module = {
     version: utilities.getVersion(),
     construct: (name: string, type: string, urn: string): pulumi.Resource => {
         switch (type) {
+            case "plant:tree/v1:Nursery":
+                return new Nursery(name, <any>undefined, { urn })
             case "plant:tree/v1:RubberTree":
                 return new RubberTree(name, <any>undefined, { urn })
             default: