+++ sdk/nodejs/x/automation/index.ts
@@ -20,3 +20,4 @@ export * from "./stackSettings";
 export * from "./projectSettings";
 export * from "./localWorkspace";
 export * from "./workspace";
+export * from "./events";